Update solution for pytorch-tensor-creation

+++ study-plans/pytorch-basics/pytorch-tensor-creation/pytorch-tensor-creation.py
@@ -6,11 +6,11 @@ def create_tensor(method, shape, value=0.0):
     """
     x = None
     if method == "zeros":
-        x = torch.zeros(shape) 
+        x = torch.zeros(shape, dtype=torch.float32) 
     elif method == "ones":
-        x = torch.ones(shape)
+        x = torch.ones(shape, dtype=torch.float32)
     elif method == "full":
-        x = torch.full(shape, value)
+        x = torch.full(shape, value, dtype=torch.float32)
     else:
         raise Exception(f"invalid method {method}")
     return x.tolist()
